+++ src/daemonclient.cpp
@@ -1,10 +1,24 @@
 #include "daemonclient.h"
 
+#include <QProcess>
+#include <QDebug>
+
 #include <sys/socket.h>
 #include <sys/un.h>
 
 constexpr sockaddr_un m_saddr = {AF_UNIX, "/tmp/optimus-manager"};
 constexpr socklen_t m_addrlen = sizeof(m_saddr);
+
+bool DaemonClient::isDaemonActive()
+{
+    QProcess process;
+    process.setProgram("systemctl");
+    process.setArguments({"is-active", "optimus-manager"});
+    process.start();
+    process.waitForFinished();
+
+    return process.readAllStandardOutput() == "active\n";
+}
 
 void DaemonClient::connect()
 {