+++ CHANGELOG.md
@@ -16,6 +16,7 @@ All notable changes to this project will be documented in this file. This projec
 -   Show application name in messages.
 -   Display an error icon if the mode icon is not found.
 -   Fix applying the tray icon after changing settings.
+-   Use optimus-manager daemon directly to switch GPU.
 
 ## 1.0.0 (2018-11-17)